requirements.txt
amqp==5.3.1
anyio==4.9.0
asgiref==3.8.1
beautifulsoup4==4.13.4
billiard==4.2.1
boto3==1.35.70
botocore==1.35.99
Brotli==1.1.0
celery==5.5.3
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
cssselect2==0.8.0
Django==5.2
django-anymail==13.0
django-celery-beat==2.8.1
django-cors-headers==4.7.0
django-markdownx==4.0.7
django-storages==1.14.6
django-timezone-field==7.1
djangorestframework==3.16.0
ecdsa==0.19.1
fonttools==4.58.4
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jmespath==1.0.1
kombu==5.5.4
lxml==6.0.0
Markdown==3.8.2
MarkupSafe==3.0.2
packaging==25.0
pillow==11.2.1
prompt_toolkit==3.0.51
psycopg2-binary==2.9.9
pycparser==2.22
pydyf==0.11.0
PyJWT==2.9.0
pyphen==0.17.2
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
python-http-client==3.3.7
pytz==2025.2
redis==6.2.0
requests==2.32.3
s3transfer==0.10.4
sendgrid==6.12.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
sqlparse==0.5.3
stripe==11.6.0
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
weasyprint==65.1
webencodings==0.5.1
Werkzeug==3.1.3
zopfli==0.2.3.post1
